Add files via upload
This is the first initial completed version of my homework tracker application. The initial idea behind this was just to assist with my homework and later I decided to also audit my homework. Initial build time was around an hour maybe a bit more. The objective for this application is to improve upon it so that it becomes a multifaceted practice tool for math and an array of other subjects.